Backend/requirements.txt
fastapi
uvicorn
python-dotenv
firecrawl-py